+++ Changelog.md
@@ -3,6 +3,7 @@
     -Changed popup.html to not use alert
     -Added popup.html Feedback
     -secret.json file handling improved
+    -Removed need of popup.html
 
 2.0.3: 27.01.2020
     -Updated README.md